Fixed mainnet issues and issues with custom network recreation on storage loads
